# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Immobilienverwaltung API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Units ───────────────────────────────────────────────────────────────────
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a tenant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Contracts ───────────────────────────────────────────────────────────────
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      parameters:
        - name: unitId
          in: query
          required: false
          schema:
            type: integer
        - name: tenantId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Documents ───────────────────────────────────────────────────────────────
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
        - name: unitId
          in: query
          required: false
          schema:
            type: integer
        - name: contractId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Utility Costs ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Property:
      type: object
      required: [id, name, address, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        type:
          type: string
          enum: [apartment_building, house, commercial, land]
        description:
          type: ["string", "null"]
        purchasePrice:
          type: ["number", "null"]
        purchaseYear:
          type: ["integer", "null"]
        totalUnits:
          type: integer
        createdAt:
          type: string
          format: date-time
    PropertyInput:
      type: object
      required: [name, address, type]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        type:
          type: string
          enum: [apartment_building, house, commercial, land]
        description:
          type: string
        purchasePrice:
          type: number
        purchaseYear:
          type: integer
        totalUnits:
          type: integer
          default: 1
    PropertyUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        type:
          type: string
          enum: [apartment_building, house, commercial, land]
        description:
          type: string
        purchasePrice:
          type: number
        purchaseYear:
          type: integer
        totalUnits:
          type: integer

    # ─── Unit ──────────────────────────────────────────────────────────────────
...
    Tenant:
      type: object
      required: [id, firstName, lastName, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TenantInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        notes:
          type: string
    TenantUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        notes:
          type: string

    # ─── Contract ──────────────────────────────────────────────────────────────
    Contract:
      type: object
      required: [id, unitId, tenantId, startDate, monthlyRent, status, createdAt]
      properties:
        id:
          type: integer
        unitId:
          type: integer
        tenantId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        monthlyRent:
          type: number
        deposit:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, terminated, pending]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ContractInput:
      type: object
      required: [unitId, tenantId, startDate, monthlyRent]
      properties:
        unitId:
          type: integer
        tenantId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        monthlyRent:
          type: number
        deposit:
          type: number
        status:
          type: string
          enum: [active, terminated, pending]
          default: active
        notes:
          type: string
    ContractUpdate:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        monthlyRent:
          type: number
        deposit:
          type: number
        status:
          type: string
          enum: [active, terminated, pending]
        notes:
          type: string

    # ─── Document ──────────────────────────────────────────────────────────────
    Document:
      type: object
      required: [id, name, fileUrl, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: ["string", "null"]
        fileUrl:
          type: string
        mimeType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        propertyId:
          type: ["integer", "null"]
        unitId:
          type: ["integer", "null"]
        contractId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── UtilityCost ───────────────────────────────────────────────────────────
...