# lib/api-zod/src/generated/types/contractUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Immobilienverwaltung API
 * OpenAPI spec version: 0.1.0
 */
import type { ContractUpdateStatus } from './contractUpdateStatus';

export interface ContractUpdate {
  startDate?: Date;
  endDate?: Date;
  monthlyRent?: number;
  deposit?: number;
  status?: ContractUpdateStatus;
  notes?: string;
}