# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Immobilienverwaltung API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyType } from './propertyType';

export interface Property {
  id: number;
  name: string;
  address: string;
  type: PropertyType;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  purchasePrice?: number | null;
  /** @nullable */
  purchaseYear?: number | null;
  totalUnits?: number;
  createdAt: Date;
}