# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Immobilienverwaltung API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdateType } from './propertyUpdateType';

export interface PropertyUpdate {
  name?: string;
  address?: string;
  type?: PropertyUpdateType;
  description?: string;
  purchasePrice?: number;
  purchaseYear?: number;
  totalUnits?: number;
}