# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Immobilienverwaltung API
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dateOfBirth?: Date | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}